requirements.txt
customtkinter>=5.2
psutil>=5.9
speedtest-cli>=2.1
plyer>=2.1